backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
httpx==0.25.2
python-magic==0.4.27
aiofiles==23.2.1
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
accelerate==0.24.0
google-generativeai==0.8.3
sentence-transformers==2.2.2
numpy>=1.21.0 